move multiplyAndFormatUnits to frontend dir
