+++ packages/utils/mod.ts
@@ -146,14 +146,6 @@ export function formatUnitsFromString(price: string, decimal: number) {
   return formatUnits(BigInt(price), decimal);
 }
 
-export function multiplyAndFormatUnits(
-  price: string,
-  quantity: number,
-  decimals: number,
-) {
-  return formatUnits(BigInt(price) * BigInt(quantity), decimals);
-}
-
 interface AdressObj {
   address: `0x${string}`;
   chainId: number;
